refactor(fleet-change): extract fleet alignment
